+++ weather@mockturtl/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.1.3
+* Update DarkSky API cutoff date
+* Update TS declarations
+* Resolved [#3926](https://github.com/linuxmint/cinnamon-spices-applets/issues/3926) Change Climacell naming to Tomorrow.io
+* Update screenshot
+
 ## 3.1.2
 
 * Do not minify code so .pot files can be generated without missing strings.